stat_arb/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
[project]
name = "stat-arb-pipeline"
version = "0.1.0"
description = "Avellaneda & Lee statistical arbitrage pipeline"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyarrow>=14.0",
  "PyYAML>=6.0",
  "pandas-market-calendars>=4.3",
  "scikit-learn>=1.3",
]
[project.optional-dependencies]
dev = ["pytest>=7.0"]
[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
